Make standard setting
